fix: sep node types
