+++ packages/server/core/src/index.ts
@@ -1,12 +1,5 @@
 export * from './core/plugin';
 export * from './core/render';
-export type {
-  HonoContext,
-  HonoEnv,
-  HonoMiddleware,
-  HonoRequest,
-} from './core/hono';
 export * from '@modern-js/plugin';
 export * from './core/loadPlugins';
 export * from './types/config';
-export type { ServerEnv, ServerManifest } from './core/server';
