+++ packages/server/plugin-express/tests/setup.ts
@@ -1,5 +1,5 @@
-jest.mock('@modern-js/server-core/base', () => {
-  const originalModule = jest.requireActual('@modern-js/server-core/base');
+jest.mock('@modern-js/server-core/base/node', () => {
+  const originalModule = jest.requireActual('@modern-js/server-core/base/node');
 
   return {
     ...originalModule,
